Feature/retrieve scxa (#19)
* Fix option ${save_raw} in <command/>

If boolean ${opt} is set in the recommended way, i.e. "--opt" when
true and "" when false, then add single quotes around, i.e. '${opt}'
will break the executed command, as argparse will treat '' as an
empty positional argument which it doesn't recognise. This fix simply
removes the double quote as a temporary measure. Better ways need to
be sought.

* Move plotting related part of the command in macros as @PLOT_OPTS@

Also change --color to --color-by to accommodate change in bioconda
wrapper.

* Add galaxy wrapper for retrieve dataset from sc expression atlas

* Add retrieve-scxa to tool_conf.xml

* Add galaxy wrapper for retrieving data from sc expression atlas

* Remove scanpy/retrieve-sca.xml

* Add help and citation to make the linter happy

* Changes tool name and adds conda requirement.