Correct method in the tool description
The method should be Gaussian not diffmap (this isn't a method, it's one of the embeddings you can use) - see the documentation on the Scanpy pp.neighbours https://scanpy.readthedocs.io/en/stable/generated/scanpy.pp.neighbors.html - which confirms that the method used in the cited papers is gauss/gaussian. 
This really confused me so it would be good to change it to prevent anyone else trying to work out why the diffmap doesn't exist as a method on the tool!